Adding xolox/vim-misc because xolox/vim-reload depends on it
Was getting the following error each time I opened vim:

Warning: The vim-reload plug-in requires the vim-misc plug-in which seems not to be installed! For more information please review the installation instructions in the readme (also available on the homepage and on GitHub). The vim-reload plug-in will now be disabled.
Press ENTER or type command to continue

This solves that. 

Cheers